# src/main/java/com/ecommerce/project/security/request/LoginRequest.java
package com.ecommerce.project.security.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;

@Setter
@Getter
@ToString
public class LoginRequest {
    @NotBlank
    private String username;

    @NotBlank
    private String password;

}